Added static code analysis
Cleaned up project a bit
