+++ scripts/run_server.py
@@ -1,3 +1,7 @@
+"""
+Script for running flask server
+"""
+
 import argparse
 import sys
 
@@ -6,17 +10,17 @@ import traffic.utilities
 
 
 def main():
+    """
+    Script entry point
+    """
 
     parser = argparse.ArgumentParser()
     parser.add_argument("--config_path", help="path to config file", required=True)
 
     config = traffic.utilities.get_yaml_configuration(sys.argv[1:])
 
-    print("Passing configuration")
-    print(config)
-
-    app = traffic.web.server.get_configured_app(configuration=config)
-    app.run(host="0.0.0.0", port=5000, debug=True)
+    app = traffic.web.server.get_configured_app(config)
+    app.run(host="0.0.0.0", port=5000)
 
 
 if __name__ == "__main__":
