+++ setup.cfg
@@ -1,0 +1,6 @@
+[pycodestyle]
+
+max-line-length = 120
+
+ignore = E402,E125,E711
+exclude = setup.cfg, *.yaml, *.md
