+++ traffic/invoke/tests.py
@@ -1,0 +1,40 @@
+"""
+Module with tests tasks for invoke
+"""
+
+import invoke
+
+
+@invoke.task
+def static_code_analysis(context):
+    """
+    Static code analysis task.
+    Runs pylint, pycodestyle, xenon and eslint
+    """
+
+    python_code_directories = "./traffic ./scripts ./tests"
+
+    context.run("pycodestyle {}".format(python_code_directories), echo=True)
+    context.run("xenon . --max-absolute=B", echo=True)
+    context.run("pylint {}".format(python_code_directories), echo=True)
+
+
+@invoke.task
+def unit_tests(context):
+    """
+    Unit tests task.
+    Runs pytest on tests defined in commit stage directory
+    """
+
+    context.run("py.test ./tests/commit_stage/unit_tests", echo=True, pty=True)
+
+
+@invoke.task
+def commit_stage(context):
+    """
+    Commit stage task.
+    Runs unit tests and static code analysis tasks
+    """
+
+    unit_tests(context)
+    static_code_analysis(context)
